requirements.txt
streamlit - 1.43.2
numpy - 2.2.4
torch -2.6.0
torchvision - 0.21.0
opencv-python(CV2) - 4.11.0.86
pandas == 2.2.3
pillow == 11.1.0
